Rename model CF to WMF
