+++ examples/wmf_example.py
@@ -26,7 +26,7 @@ data = citeulike.load_data(reader=Reader(item_set=item_ids))
 ratio_split = RatioSplit(data=data, test_size=0.2, exclude_unknowns=True,
                          verbose=True, seed=123, rating_threshold=0.5)
 
-cf = cornac.models.CF(k=50, max_iter=50, learning_rate=0.001, lambda_u=0.01, lambda_v=0.01, verbose=True, seed=123)
+cf = cornac.models.WMF(k=50, max_iter=50, learning_rate=0.001, lambda_u=0.01, lambda_v=0.01, verbose=True, seed=123)
 
 rec_300 = cornac.metrics.Recall(k=300)
